GCS_MAVLink: move MissionItemProtocol functions out of GCS_Common
